API

Index

GFit.CartesianDomain

GFit.Domain

GFit.FitResult

GFit.Measures

GFit.Model

GFit.MultiModel

GFit.Parameter

GFit.λFunct

Base.getindex

Base.haskey

Base.length

Base.push!

Base.values

GFit.axis

GFit.coords

GFit.domain

GFit.evaluate

GFit.fit!

GFit.freeze!

GFit.isfreezed

GFit.mock

GFit.select_maincomp!

GFit.thaw!

GFit.uncerts

GFit.@λ

Exported symbols

The list of GFit.jl exported symbols is as follows:

GFit.CartesianDomain — Type

CartesianDomain{N}

An object representing a model, or a dataset, N-dimensional cartesian domain (i.e. a grid).

Available constructors:

Domain(axis...): each argument is a vector containing the coordinates on a given axis (arguments may have different lengths);

Domain(lengths...): returns a N-dim CartesianDomain object whose axis lengths are specified in the arguments.

Note that a CartesianDomain requires at least 2 dimensions.

Coordinates for all points along a given axis can be obtained with the coords() function, while the coordinates of the grid can be obtained with axis().

GFit.Domain — Type

Domain{N}

An object representing a model, or a dataset, N-dimensional linear domain.

Available constructors:

Domain(coords...): each argument is a vector, one for each dimension (all arguments must have same lengths);

Domain(length): returns a 1-dim Domain object of the given length.

Coordinates for all points along a given axis can be obtained with the coords function.

GFit.Measures — Type

Measures{N}

An object representing a set of empirical measurements (with Gaussian uncertainties)

Available constructors:

Measures(domain::Domain{N}, values::AbstractVector{T}, uncerts::AbstractVector{T}) where {T <: AbstractFloat, N}

Measures(domain::CartesianDomain{N}, values::AbstractArray{T, N}, uncerts::AbstractArray{T, N}) where {T <: AbstractFloat, N}

In the above constructor methods the last argument may also be a scalar value, to set the same uncertainty for all the measurements. The method accepting a CartesianDomain requires arrays with at least 2 dimensions.

The domain, values and uncertainties for a Measures object can be retrieved using the domain, values and uncerts functions respectively.

GFit.Model — Type

Model

A structure containing a model description, whose evaluation is suitable to be compared to a single empirical dataset.

Constructor is: Model(domain::AbstractDomain, components...) where the first argument is either a Domain or CartesianDomain object, and the remaining one(s) is (are) the model component(s), which may be given as:

a single Dict{Symbol, AbstractComponent}, where the keys are the names and the values the component objects;

a single component, which will have a default name is assigned (:main);

a single λFunct, which will be wrapped into an LComp component and a default name will be assigned (:main);

one or more Pair{Symbol, AbstractComponent}, where the first element is the name and the second is the component.

You may access the individual component in a Model using the indexing syntax, as if it was a Dict{Symbol, AbstractComponent}. Also, you may add new components to a Model after it has been created using the same synatx. Finally, you may use the keys() and haskey() functions with their usual meanings.

Individual components may be freezed (i.e. have all its parameters fixed during fitting, despite the individual Parameter settings) or thawed using the freeze!() and thaw!() functions. Use the isfreezed() function to check if a component is freezed.

The main component, i.e. the one whose evaluation corresponds to the overall model evaluation, is typically automatically identified by analyzing the component dependencies. However a specific component may be forced to be the main one by invoking select_maincomp!.

The model is automatically evaluated whenever needed, however there are a few cases where it is not possible to trigger an automatic evaluation, e.g. immediately after the user modifies a Parameter value. In this case an evaluation can be forced by invoking evaluate().

The most important function for a Model object is fit!(), which allows to fit the model against an empirical dataset. The ! in the name reminds us that, after fitting, the parameter values will be set to the best fit ones (rather than retaining their original values).

The model and all component evaluation can be obtained by using the Model object has if it was a function: with no arguments it will return the main component evaluation, while if a Symbol is given as argument it will return the evaluation of the component with the same name.

GFit.MultiModel — Type

MultiModel

A structure containing a multi-model description, whose evaluation is suitable to be compared to a set of empirical datasets. A MultiModel is very similar to a vector of Model objects, with the ability to trasparently handle the patch constraints between one model and the other.

Constructor is: MultiModel(model1, model2, ...).

You may access the individual Model objects the indexing syntax, as if it was a Vector{Model}. Also, you may add new model to a MultiModel after it has been created using the push!() function. Finally, you may retrieve the length of the vector with length().

Just like a Model object you may need to manually trigger a MultiModel evaluation using the evaluate() function.

GFit.@λ — Macro

@λ expr

Macro to generate λFunct objects using the same syntax as in a standard Julia anonymous function. Example: @λ (x, f) -> f .* x

GFit.axis — Function

axis(d::CartesianDomain, dim::Integer)

Returns the coordinates of the grid along a given dimension as a Vector{Float64}.

GFit.coords — Function

coords(d::Domain{1})
coords(d::Domain, dim::Integer)
coords(d::CartesianDomain, dim::Integer)

Returns coordinates of all points along a given dimension as a Vector{Float64}.

GFit.domain — Function

domain(model::Model)

Return the domain where the model is evaluated.

GFit.evaluate — Function

evaluate(model::Model)

Evaluate a Model and update internal structures.

evaluate(multi::MultiModel)

Evaluate a MultiModel and update internal structures.

GFit.fit! — Function

fit!(model::Model, data::Measures; minimizer::AbstractMinimizer=lsqfit())

Fit a model to an empirical data set using the specified minimizer (default: lsqfit()).

fit!(multi::MultiModel, data::Vector{Measures{N}}; minimizer::AbstractMinimizer=lsqfit())

Fit a multi-model to a set of empirical data sets using the specified minimizer (default: lsqfit()).

GFit.freeze! — Function

freeze!(model::Model, cname::Symbol)

Freeze a component in the model (i.e. treat all component parameters as fixed for fitting).

Base.getindex — Function

getindex(d::Union{Domain, CartesianDomain}, dim::Integer)

Shortcut for coords(d, dim).

Base.haskey — Function

haskey(m::Model, name::Symbol)

Check whether a component exists in model.

GFit.isfreezed — Function

isfreezed(model::Model, cname::Symbol)

Check whether a component is freezed in the model.

Base.length — Function

length(multi::MultiModel)

Returns how many Model objects are contained in a MultiModel.

Base.push! — Function

push!(multi::MultiModel, model::Model)

Push a new Model object into a MultiModel.

GFit.select_maincomp! — Function

select_maincomp!(model::Model, cname::Symbol)

Force a component to be the final one for model evaluation.

GFit.thaw! — Function

thaw!(model::Model, cname::Symbol)

Thaw a freezed component in the model (i.e. treat component parameters as fixed only if explicitly set in the corresponding Parameter structure).

GFit.uncerts — Function

uncerts(d::Measures)

Returns the measurement uncertainties as a Vector{Float64}.

Base.values — Function

values(d::Measures)

Returns the measurement values as a Vector{Float64}.

Non-exported symbols

The following symbols are not exported by the GFit.jl package since they are typically not used in every day work, or aimed to debugging purposes. Still, they can be useful in some case, hence they are documented here.

GFit.FitResult — Type

FitResult

A structure representing the results of a fitting process.

Fields:

timestamp::DateTime: time at which the fitting process has started;

elapsed::Float64: elapsed time (in seconds);

ndata::Int: number of data empirical points;

nfree::Int: number of free parameters;

dof::Int: ndata - nfree;

fitstat::Float64: fit statistics (equivalent ro reduced χ^2 for Measures objects);

status: minimizer exit status (tells whether convergence criterion has been satisfied, or if an error has occurred during fitting);

bestfit: dictionary of Parameter objects containing best fit values and their 1-σ uncertainties. In the multi-model case this will be a vector of dictionaries.

Note: the FitResult fields are supposed to be accessed directly by the user, without invoking any get/set method.

GFit.λFunct — Type

λFunct

A representation for a λ-function, containing a reference to the function itself, a string representation of its source code (for displaying purposes) and a list of arguments. It is typically created using the @λ macro.

GFit.Parameter — Type

Parameter

A structure representing a model parameter.

Fields:

val::Float64: parameter value (initial guess before fitting, or best fit one after fitting);

low::Float64: lower limit for the value (default: -Inf);

high::Float64: upper limit for the value (default: +Inf);

fixed::Bool: whether the parameter is fixed during fitting (default: false);

patch::Union{Nothing, Symbol, λFunct}: patch prescription within the same model;

mpatch::Union{Nothing, λFunct}: patch prescription in a multi-model analysis;

actual::Float64: actual value for the parameter (i.e. after applying the patch prescription)`;

unc::Float64: 1σ uncertainty associated to the parameter value.

Note: the Parameter fields are supposed to be accessed directly by the user, without invoking any get/set method.

GFit.mock — Function

mock(::Type{Measures}, model::Model; keywords...)
mock(::Type{Measures}, multi::MultiModel; keywords...)

Generate mock dataset(s) using a ground truth Model or MultiModel object. The first version returns a single Measures object, while the second returns a Vector{Measures}.

The measurement random errors added to the data points are drawn from a Normal distribution centered on the data value itself, and a width given by the sum of three contributions:

proportional part: error proportional to each data point value;

range part: error proportional to the range spanned by all values in a single dataset;

absolute part: absolute error value.

No systematic error is considered when generating mock dataset(s).

Accepted keywords:

properr=0.01: proportional error;

rangeerr=0.05: range error;

abserr=0.: absolute error;

seed=nothing: seed for the Random.MersenneTwister generator.